Move predictor to ray.train
