+++ python/ray/air/__init__.py
@@ -2,7 +2,7 @@ from ray.air.checkpoint import Checkpoint
 from ray.air.data_batch_type import DataBatchType
 from ray.air.config import RunConfig, ScalingConfig, DatasetConfig
 from ray.data.preprocessor import Preprocessor
-from ray.air.predictor import Predictor
+from ray.train.predictor import Predictor
 from ray.air.result import Result
 from ray.air.batch_predictor import BatchPredictor
 from ray.air.util.datasets import train_test_split
