+++ python/ray/air/batch_predictor.py
@@ -2,7 +2,7 @@ from typing import Type, Optional, Dict, Any, Union
 
 import ray
 from ray.air import Checkpoint
-from ray.air.predictor import Predictor
+from ray.train.predictor import Predictor
 from ray.air.util.data_batch_conversion import convert_batch_type_to_pandas
 from ray.util.annotations import PublicAPI
