+++ python/ray/air/predictors/integrations/lightgbm/lightgbm_predictor.py
@@ -5,7 +5,7 @@ import pandas as pd
 
 from ray.air.checkpoint import Checkpoint
 from ray.air.constants import TENSOR_COLUMN_NAME
-from ray.air.predictor import Predictor
+from ray.train.predictor import Predictor
 from ray.train.lightgbm import load_checkpoint
 
 if TYPE_CHECKING:
