+++ python/ray/air/predictors/integrations/torch/torch_predictor.py
@@ -4,7 +4,7 @@ import numpy as np
 import pandas as pd
 import torch
 
-from ray.air.predictor import Predictor, DataBatchType
+from ray.train.predictor import Predictor, DataBatchType
 from ray.air.checkpoint import Checkpoint
 from ray.train.torch import load_checkpoint
 from ray.air._internal.torch_utils import convert_pandas_to_torch_tensor
