+++ python/ray/air/predictors/integrations/xgboost/xgboost_predictor.py
@@ -5,7 +5,7 @@ import xgboost
 
 from ray.air.checkpoint import Checkpoint
 from ray.air.constants import TENSOR_COLUMN_NAME
-from ray.air.predictor import Predictor
+from ray.train.predictor import Predictor
 from ray.train.xgboost import load_checkpoint
 
 if TYPE_CHECKING:
